Fantasy changes and added some pages
Now easier to tell if your fantasy lineup is saved and if you are making changes to it or not. Current lineup now displays on team load. Made pages for create and join league but nothing inside the pages yet. Sidebar now dropsdown when you hover over fantasy and links to these pages shows.